drf/requirements.txt
algoliasearch-django>=4.0,<5.0
django>=4.0.0
djangorestframework
pyyaml
requests
django-cors-headers